Include gradient direction in cost function
This commit contains updates to the cost function, but the rest of the
system needs to be updated to accept the new cost function.

The new cost function contains the cost for each pixel as well as the
cost for traveling from each pixel to each adjacent pixel (including
diagonals). So the cost function has twice the dimensions of the image in
question, and some translation will need to be performed from cost
function coordinates to image coordinates.